requirements.txt
streamlit
google-generativeai
langchain-community
psycopg2
sqlalchemy
langchain
python-dotenv